changed real/float to numeric/decimal for money values

+++ lib/sanbase/model/latest_btc_wallet_data.ex
@@ -5,7 +5,7 @@ defmodule Sanbase.Model.LatestBtcWalletData do
 
   schema "latest_btc_wallet_data" do
     field :address, :string
-    field :satoshi_balance, :float
+    field :satoshi_balance, :decimal
     field :update_time, Timex.Ecto.DateTime
   end
